python move to partition.py
