+++ python/rapidsmpf/rapidsmpf/cudf/CMakeLists.txt
@@ -3,18 +3,10 @@
 # SPDX-License-Identifier: Apache-2.0
 # =================================================================================
 
-set(cython_modules config.pyx progress_thread.pyx rmm_resource_adaptor.pyx shuffler.pyx
-                   statistics.pyx
-)
+set(cython_modules partition.pyx)
 
 rapids_cython_create_modules(
   CXX
   SOURCE_FILES "${cython_modules}"
-  LINKED_LIBRARIES rapidsmpf::rapidsmpf
+  LINKED_LIBRARIES rapidsmpf::rapidsmpf ucxx::ucxx ucx::ucp $<TARGET_NAME_IF_EXISTS:MPI::MPI_C>
 )
-
-add_subdirectory(_detail)
-add_subdirectory(buffer)
-add_subdirectory(communicator)
-add_subdirectory(cudf)
-add_subdirectory(utils)
