+++ python/rapidsmpf/rapidsmpf/cudf/__init__.pxd
@@ -1,0 +1,2 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES.
+# SPDX-License-Identifier: Apache-2.0
