+++ python/rapidsmpf/rapidsmpf/shuffler.pyi
@@ -2,10 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 from __future__ import annotations
 
-from collections.abc import Iterable, Mapping
+from collections.abc import Mapping
 
-from pylibcudf.table import Table
-from rmm.pylibrmm.memory_resource import DeviceMemoryResource
 from rmm.pylibrmm.stream import Stream
 
 from rapidsmpf.buffer.packed_data import PackedData
@@ -13,25 +11,6 @@ from rapidsmpf.buffer.resource import BufferResource
 from rapidsmpf.communicator.communicator import Communicator
 from rapidsmpf.progress_thread import ProgressThread
 from rapidsmpf.statistics import Statistics
-
-def partition_and_pack(
-    table: Table,
-    columns_to_hash: Iterable[int],
-    num_partitions: int,
-    stream: Stream,
-    device_mr: DeviceMemoryResource,
-) -> dict[int, PackedData]: ...
-def split_and_pack(
-    table: Table,
-    splits: Iterable[int],
-    stream: Stream,
-    device_mr: DeviceMemoryResource,
-) -> dict[int, PackedData]: ...
-def unpack_and_concat(
-    partitions: Iterable[PackedData],
-    stream: Stream,
-    device_mr: DeviceMemoryResource,
-) -> Table: ...
 
 class Shuffler:
     max_concurrent_shuffles: int
